Challenge 1,2,3,4
The First challenges completed, I think. I did not quite understand what to do for the 2 Structuring Conent. I was hoping to get some more information on that during the next seminar.

Also, I left the HTML Terms the way they were, because I likes the look of that better than putting it in a <table>.

Hope I manage to hand it in, because GitHub confuses me. B)

Richie Elia, UXD 1A 18055354